% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsurv.R
\name{plotsurv}
\alias{plotsurv}
\title{plotsurv}
\usage{
plotsurv(survfit_obj,
         include_surv = TRUE,
         include_CIFs = NULL,
         display_event = "all",
         title = "Cumulative incidence and survival",
         subtitle = "",
         x_lab = "Time",
         y_lab = "Cumulative probability of event",
         color_lab = "Group and event type",
         fill_lab = "Group and event type",
         line_lab = "Group and event type",
         group_labels = NULL,
         colors = NULL,
         linetypes = NULL
       )
}
\arguments{
\item{survfit_obj}{the output of a survift() call}

\item{include_surv}{(optional) a character value for the document title}

\item{include_CIFs}{(optional) a character value for the document subtitle}

\item{display_event}{(optional) a vector of event values to be displayed (defaults to "all"),}

\item{title}{(optional) The plot title, defaults to "Cumulative incidence and survival"}

\item{subtitle}{(optional) The plot subtitle, defaults to empty string ""}

\item{x_lab}{(optional) X-axis label, defaults to "Time"}

\item{y_lab}{(optional) Y-axis label, defaults to "Cumulative probability of event"}

\item{color_lab}{(optional) Label for color-codes, defaults to "Group and event type" (if same as fill_lab and line_lab, one combined legend will apear)}

\item{fill_lab}{(optional) Label for color-fills, defaults to "Group and event type" (if same as color_lab and line_lab, one combined legend will apear)}

\item{line_lab}{(optional) Label for lines, defaults to "Group and event type" (if same as color_lab and fill_lab, one combined legend will apear)}

\item{group_labels}{(optional) Labels for each subgroup (defaults to automated combination of RHS group and LHS event type), must be of same length as no. of groups}

\item{colors}{(optional) Colors for each subgroup, must be of same length as no. of groups}

\item{linetypes}{(optional) Linetypes for each subgroup, must be of same length as no. of groups}
}
\value{
ggplot object
}
\description{
Create nice plots for Kaplan-Meier/Aalen-Johansen models fitted using survival::survfit().
}
\examples{
  \dontrun{
  plotsurv(
    data = fit_obj,
    include_surv = FALSE,
    title = "Cumulative incidence of metastatic breast cancer recurrence",
    subtitle = "Between 2015 and 2023",
    colors = c("#3A488A","#BD5630")
  )
  }
}
\author{
Zheer Kejlberg Al-Mashhadi
}
